rename dir (#3562)
* rename and move

* rename test dir
