--- cg/services/post_processing/pacbio/run_data_generator/run_data.py
@@ -1,12 +1,0 @@
-from pathlib import Path
-
-from cg.services.post_processing.abstract_models import RunData
-
-
-class PacBioRunData(RunData):
-    """Holds information on a single SMRTcell of a PacBio run."""
-
-    full_path: Path
-    sequencing_run_name: str
-    well_name: str
-    plate: int
